Phase1_PilotProject/requirements.txt
flask
werkzeug
requests
docx
pyenchant	
PyPDF2